refactor code
refactoring

+++ MultiThreading.Task3.Matrixes/Multipliers/MatricesMultiplier.cs
@@ -9,10 +9,10 @@ namespace MultiThreading.Task3.MatrixMultiplier.Multipliers
             var resultMatrix = new Matrix(m1.RowCount, m2.ColCount);
             for (long i = 0; i < m1.RowCount; i++)
             {
-                for (byte j = 0; j < m2.ColCount; j++)
+                for (long j = 0; j < m2.ColCount; j++)
                 {
                     long sum = 0;
-                    for (byte k = 0; k < m1.ColCount; k++)
+                    for (long k = 0; k < m1.ColCount; k++)
                     {
                         sum += m1.GetElement(i, k) * m2.GetElement(k, j);
                     }
